Fix events

+++ ReportCore/src/main/java/com/operators/reportrejectcore/ReportCore.java
@@ -88,7 +88,7 @@ public class ReportCore {
         }
     }
 
-    public void sendStopReport(int stopReasonId, int stopSubReasonId, int eventId, int jobId) {
+    public void sendStopReport(int stopReasonId, int stopSubReasonId, long eventId, int jobId) {
         if (mReportPersistenceManagerInterface != null) {
             mReportRejectNetworkBridgeInterface.sendReportStop(mReportPersistenceManagerInterface.getSiteUrl(), mReportPersistenceManagerInterface.getSessionId(),
                     String.valueOf(mReportPersistenceManagerInterface.getMachineId())
