+++ ReportNetworkBridge/src/main/java/com/operators/reportrejectnetworkbridge/ReportNetworkBridge.java
@@ -61,7 +61,7 @@ public class ReportNetworkBridge implements ReportRejectNetworkBridgeInterface {
 
 
     @Override
-    public void sendReportStop(String siteUrl, String sessionId, String machineId, String operatorId, int stopReasonId, int stopSubReasonId, int eventId, Integer jobId, final SendReportStopCallback callback, final int totalRetries, int specificRequestTimeout) {
+    public void sendReportStop(String siteUrl, String sessionId, String machineId, String operatorId, int stopReasonId, int stopSubReasonId, long eventId, Integer jobId, final SendReportStopCallback callback, final int totalRetries, int specificRequestTimeout) {
 
         SendReportStopRequest sendReportStopRequest = new SendReportStopRequest(sessionId, machineId, operatorId, stopReasonId, stopSubReasonId, jobId, eventId);
